♻️ AppLayout 파일을 삭제하고 AppRoot에 병합 (#46)
AppLayout 파일을 삭제하고 AppRoot에 병합하였습니다.
불필요한 파일을 늘려서 depth를 복잡하게 하기 때문입니다.
