+++ src/routes/app/AppRoot.tsx
@@ -1,7 +1,11 @@
 import { useEffect } from "react";
-import { Outlet, useNavigate } from "react-router-dom";
+import { Outlet, useNavigate, redirect } from "react-router-dom";
+import styled from "@emotion/styled";
 
-import AppLayout from "@routes/app/components/AppLayout";
+import Header from "@components/Header";
+import Sidebar from "@components/Sidebar";
+
+import authService from "@services/auth.service";
 import useTodoList from "@hooks/useTodoList";
 
 function AppRoot() {
@@ -12,11 +16,51 @@ function AppRoot() {
     if (!user) navigate("/");
   }, [navigate, user]);
 
+  const handleClickLogout = () => {
+    authService.logOut();
+    redirect("/");
+  };
+
   return (
-    <AppLayout>
-      <Outlet />
-    </AppLayout>
+    <Container>
+      <Header>
+        <LogOutButton onClick={handleClickLogout}>로그아웃</LogOutButton>
+      </Header>
+      <Main>
+        <Sidebar />
+        <Content>
+          <Outlet />
+        </Content>
+      </Main>
+      <Footer />
+    </Container>
   );
 }
 
 export default AppRoot;
+
+const Container = styled.div`
+  height: 100%;
+  overflow: hidden;
+`;
+const Main = styled.div`
+  display: flex;
+  height: 80%;
+`;
+const LogOutButton = styled.button`
+  border: none;
+  background-color: transparent;
+  font-size: 1rem;
+  color: ${({ theme }) => theme.color.dark};
+  cursor: pointer;
+`;
+const Content = styled.div`
+  width: 100%;
+  border-radius: 1rem;
+  padding: 1rem;
+`;
+const Footer = styled.footer`
+  height: 10%;
+  background-color: ${({ theme }) => theme.color.secondary};
+  border-radius: 1rem 1rem 0 0;
+`;
